mmc: Try fixup sd hotplug hang by forcing a reset
Two updates in this commit:

1. First is to remove the EIO early failure on init, since
that will keep the device from ever showing up if its busy
initially (which is all the time w/ 4.1).

2. Second, when command preparation times out due to the card
being busy, the abort added in the previous patch can cause
problems with the state machine, since we don't propagate the
fact that we didn't send the command. This causes the command
request logic to get stuck waiting on a completion that will
never change.

So since we don't have a good mechanism to propagate the error
here, reset the host. Which seems to work. The only gotcha is
that the cur_slot may not yet be initialized, so set it before
calling reset which might otherwise cause a null pointer deref.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
(cherry picked from commit abb00460f867240b40698cf32e2cb376c04a55a8)
(adjust dw_mci_wait_while_busy() to fix build error)
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>

Conflicts:
	drivers/mmc/core/core.c
	drivers/mmc/host/dw_mmc.c

Pouet