mmc: dw_mmc-k3: Enable CAP_MMC_CMD23 on microSD slot for RPMB detection
By adding CAP_MMC_CMD23 to the capabilities of the MMC controller
that is connected to the microSD slot, this commit makes it possible
to access the RPMB partition of an eMMC chip connected to the microSD
slot (such as HardKernel's eMMC modules).

Tested with an 8GB MMC 5.0 chip (Toshiba THGBMBG6D1KBAIL).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>