tee: fix unbalanced context refcount in register shm from fd
Successful registration of a memory reference in the scope of a
TEE content must increase the context refcount. This change
adds this missing refcount increase.

The context refcount is already decremented when such shm reference
is freed by its owner, in tee_shm_release(), hence current unbalance
refcount before this path is applied.

Fixes: 02b91c70ea3a ("tee: new ioctl to a register tee_shm from a dmabuf file descriptor")
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Tested-by: Etienne Carriere <etienne.carriere@linaro.org> (Qemu armv7/v8)
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
[jf: update SHA-1 of commit in Fixes: tag after rebasing onto v4.18]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>